+++ lib/onlyoffice_mysql_helper/mysql_helper.rb
@@ -15,8 +15,10 @@ module OnlyofficeMysqlHelper
                    user: SQL_SERVER_USER_LOCAL,
                    password: SQL_SERVER_PASSWORD_LOCAL)
       @connection = Mysql2::Client.new(host: address,
+                                       port: default_mysql_port,
                                        username: user,
-                                       password: password, database: database)
+                                       password: password,
+                                       database: database)
       @database = database
     end
 
@@ -73,6 +75,13 @@ module OnlyofficeMysqlHelper
         query_values += "'#{value.to_s.delete("'")}',"
       end
       query_values.chop!
+    end
+
+    # @return [Integer] default mysql port
+    def default_mysql_port
+      return ENV['DB_PORT'] if ENV['DB_PORT']
+
+      3306
     end
   end
 end